split out tests to new file
